chore: Small refactor of DflyInstance
* Move to its own file
* Unify self.args and self.params.args earlier so it can be inspected.
